requirements.txt
appdirs==1.4.3
aspy.yaml==1.1.1
attrs==18.1.0
black==18.6b4
boto3==1.7.70
botocore==1.10.70
cached-property==1.4.3
cfgv==1.1.0
click==6.7
docutils==0.14
identify==1.1.4
jmespath==0.9.3
nodeenv==1.3.2
pre-commit==1.10.4
python-dateutil==2.7.3
PyYAML==3.13
s3transfer==0.1.13
six==1.11.0
toml==0.9.4
virtualenv==16.0.0
